add user security tab for sysadmin

+++ internal/adapter/repository/glpi/get_user_groups.go
@@ -6,7 +6,7 @@ import (
 	"github.com/saygik/go-userinfo/internal/entity"
 )
 
-func (r *Repository) GetUserGroups(id int64) (groups []entity.IdName, err error) {
+func (r *Repository) GetUserGroups(id int) (groups []entity.IdName, err error) {
 	sql := fmt.Sprintf(
 		`SELECT glpi_groups_users.groups_id AS id, glpi_groups.name AS name
 		from glpi_groups_users INNER JOIN glpi_groups ON glpi_groups.id=glpi_groups_users.groups_id
